style dictionary utility reorganization
Moved the refactored SD utilities to style-dictionary/plugins to match project structure. Updated paths in SD config to import from new location.

Updated sizing.json to pick up --select-border-width-error: 2px and regened tokens.
