+++ packages/cfpb-design-system/src/elements/abstracts/sizing.css
@@ -6,4 +6,5 @@
   --font-adjust-base: 0.517;
   --font-adjust-step: 0.0054;
   --select-border-width-default: 1px;
+  --select-border-width-error: 2px;
 }
