+++ style-dictionary/plugins/file-discovery.js
