+++ style-dictionary/plugins/platforms.js
